leopard_pontdesarts/requirements.txt
crewai
langchain
langchain_community
requests
pydantic
python-dotenv
pyyaml
duckduckgo-search
pygame